Fix exports, fully builds
