+++ cpp/src/arrow/adapters/hiveserver2/hiveserver2-test.cc
@@ -17,16 +17,183 @@
 
 #include "arrow/adapters/hiveserver2/operation.h"
 
-#include <gtest/gtest.h>
 #include <memory>
 #include <sstream>
+#include <string>
+#include <vector>
+
+#include <gtest/gtest.h>
 
 #include "arrow/adapters/hiveserver2/service.h"
 #include "arrow/adapters/hiveserver2/session.h"
-#include "arrow/adapters/hiveserver2/test-util.h"
+#include "arrow/adapters/hiveserver2/thrift-internal.h"
 
-using namespace arrow::hiveserver2;
-using namespace std;
+#include "arrow/status.h"
+#include "arrow/test-util.h"
+
+using std::string;
+using std::unique_ptr;
+using std::vector;
+
+namespace arrow {
+namespace hiveserver2 {
+
+// Convenience functions for finding a row of values given several columns.
+template <typename VType, typename CType>
+bool FindRow(VType value, CType* column) {
+  for (int i = 0; i < column->length(); ++i) {
+    if (column->data()[i] == value) {
+      return true;
+    }
+  }
+  return false;
+}
+
+template <typename V1Type, typename V2Type, typename C1Type, typename C2Type>
+bool FindRow(V1Type value1, V2Type value2, C1Type* column1, C2Type* column2) {
+  EXPECT_EQ(column1->length(), column2->length());
+  for (int i = 0; i < column1->length(); ++i) {
+    if (column1->data()[i] == value1 && column2->data()[i] == value2) {
+      return true;
+    }
+  }
+  return false;
+}
+
+template <typename V1Type, typename V2Type, typename V3Type, typename C1Type,
+          typename C2Type, typename C3Type>
+bool FindRow(V1Type value1, V2Type value2, V3Type value3, C1Type* column1,
+             C2Type* column2, C3Type column3) {
+  EXPECT_EQ(column1->length(), column2->length());
+  EXPECT_EQ(column1->length(), column3->length());
+  for (int i = 0; i < column1->length(); ++i) {
+    if (column1->data()[i] == value1 && column2->data()[i] == value2 &&
+        column3->data()[i] == value3) {
+      return true;
+    }
+  }
+  return false;
+}
+
+// Waits for this operation to reach the given state, sleeping for sleep microseconds
+// between checks, and failing after max_retries checks.
+Status Wait(const std::unique_ptr<Operation>& op,
+            Operation::State state = Operation::State::FINISHED, int sleep_us = 10000,
+            int max_retries = 100) {
+  int retries = 0;
+  Operation::State op_state;
+  RETURN_NOT_OK(op->GetState(&op_state));
+  while (op_state != state && retries < max_retries) {
+    usleep(sleep_us);
+    RETURN_NOT_OK(op->GetState(&op_state));
+    ++retries;
+  }
+
+  if (op_state == state) {
+    return Status::OK();
+  } else {
+    std::stringstream ss;
+    ss << "Failed to reach state '" << OperationStateToString(state) << "' after "
+       << retries << " retries.";
+    return Status::IOError(ss.str());
+  }
+}
+
+// Creates a service, session, and database for use in tests.
+class HS2ClientTest : public ::testing::Test {
+ protected:
+  virtual void SetUp() {
+    int conn_timeout = 0;
+    ProtocolVersion protocol_version = ProtocolVersion::PROTOCOL_V7;
+    EXPECT_OK(
+        Service::Connect(hostname, port, conn_timeout, protocol_version, &service_));
+
+    std::string user = "user";
+    HS2ClientConfig config;
+    EXPECT_OK(service_->OpenSession(user, config, &session_));
+
+    std::unique_ptr<Operation> drop_db_op;
+    EXPECT_OK(session_->ExecuteStatement(
+        "drop database if exists " + TEST_DB + " cascade", &drop_db_op));
+    EXPECT_OK(drop_db_op->Close());
+
+    std::unique_ptr<Operation> create_db_op;
+    EXPECT_OK(session_->ExecuteStatement("create database " + TEST_DB, &create_db_op));
+    EXPECT_OK(create_db_op->Close());
+
+    std::unique_ptr<Operation> use_db_op;
+    EXPECT_OK(session_->ExecuteStatement("use " + TEST_DB, &use_db_op));
+    EXPECT_OK(use_db_op->Close());
+  }
+
+  virtual void TearDown() {
+    std::unique_ptr<Operation> use_db_op;
+    EXPECT_OK(session_->ExecuteStatement("use default", &use_db_op));
+    EXPECT_OK(use_db_op->Close());
+
+    std::unique_ptr<Operation> drop_db_op;
+    EXPECT_OK(
+        session_->ExecuteStatement("drop database " + TEST_DB + " cascade", &drop_db_op));
+    EXPECT_OK(drop_db_op->Close());
+
+    EXPECT_OK(session_->Close());
+    EXPECT_OK(service_->Close());
+  }
+
+  void CreateTestTable() {
+    std::unique_ptr<Operation> create_table_op;
+    EXPECT_OK(session_->ExecuteStatement(
+        "create table " + TEST_TBL + " (" + TEST_COL1 + " int, " + TEST_COL2 + " string)",
+        &create_table_op));
+    EXPECT_OK(create_table_op->Close());
+  }
+
+  void InsertIntoTestTable(std::vector<int> int_col_data,
+                           std::vector<std::string> string_col_data) {
+    EXPECT_EQ(int_col_data.size(), string_col_data.size());
+
+    std::stringstream query;
+    query << "insert into " << TEST_TBL << " VALUES ";
+    for (size_t i = 0; i < int_col_data.size(); i++) {
+      if (int_col_data[i] == NULL_INT_VALUE) {
+        query << " (NULL, ";
+      } else {
+        query << " (" << int_col_data[i] << ", ";
+      }
+
+      if (string_col_data[i] == "NULL") {
+        query << "NULL)";
+      } else {
+        query << "'" << string_col_data[i] << "')";
+      }
+
+      if (i != int_col_data.size() - 1) {
+        query << ", ";
+      }
+    }
+
+    std::unique_ptr<Operation> insert_op;
+    EXPECT_OK(session_->ExecuteStatement(query.str(), &insert_op));
+    EXPECT_OK(Wait(insert_op));
+    Operation::State insert_op_state;
+    EXPECT_OK(insert_op->GetState(&insert_op_state));
+    EXPECT_EQ(insert_op_state, Operation::State::FINISHED);
+    EXPECT_OK(insert_op->Close());
+  }
+
+  const std::string hostname = "localhost";
+  int port = 21050;
+
+  const std::string TEST_DB = "hs2client_test_db";
+  const std::string TEST_TBL = "hs2client_test_table";
+  const std::string TEST_COL1 = "int_col";
+  const std::string TEST_COL2 = "string_col";
+
+  const int NULL_INT_VALUE = -1;
+
+  std::unique_ptr<Service> service_;
+  std::unique_ptr<Session> session_;
+};
 
 class OperationTest : public HS2ClientTest {};
 
@@ -41,7 +208,8 @@ TEST_F(OperationTest, TestFetch) {
   unique_ptr<ColumnarRowSet> results;
   bool has_more_rows = false;
   // Impala only supports NEXT and FIRST.
-  EXPECT_ERROR(select_op->Fetch(2, FetchOrientation::LAST, &results, &has_more_rows));
+  ASSERT_RAISES(IOError,
+                select_op->Fetch(2, FetchOrientation::LAST, &results, &has_more_rows));
 
   // Fetch the results in two batches by passing max_rows to Fetch.
   EXPECT_OK(select_op->Fetch(2, FetchOrientation::NEXT, &results, &has_more_rows));
@@ -112,7 +280,7 @@ TEST_F(OperationTest, TestCancel) {
 
   unique_ptr<Operation> op;
   EXPECT_OK(session_->ExecuteStatement("select count(*) from " + TEST_TBL, &op));
-  op->Cancel();
+  EXPECT_OK(op->Cancel());
   // Impala currently returns ERROR and not CANCELED for canceled queries
   // due to the use of beeswax states, which don't support a canceled state.
   EXPECT_OK(Wait(op, Operation::State::ERROR));
@@ -131,7 +299,7 @@ TEST_F(OperationTest, TestGetLog) {
   EXPECT_OK(session_->ExecuteStatement("select count(*) from " + TEST_TBL, &op));
   string log;
   EXPECT_OK(op->GetLog(&log));
-  EXPECT_TRUE(log != "");
+  EXPECT_NE(log, "");
 
   EXPECT_OK(op->Close());
 }
@@ -155,7 +323,7 @@ TEST_F(OperationTest, TestGetResultSetMetadata) {
   unique_ptr<Operation> select_op;
   EXPECT_OK(session_->ExecuteStatement("select * from " + TEST_TBL, &select_op));
   vector<ColumnDesc> column_descs;
-  select_op->GetResultSetMetadata(&column_descs);
+  EXPECT_OK(select_op->GetResultSetMetadata(&column_descs));
   EXPECT_EQ(column_descs.size(), 3);
 
   EXPECT_EQ(column_descs[0].column_name(), TEST_COL1);
@@ -222,8 +390,8 @@ TEST_F(SessionTest, TestSessionConfig) {
 
   // Ensure the we can't access the table.
   unique_ptr<Operation> select_op_error;
-  EXPECT_ERROR(
-      session_error->ExecuteStatement("select * from " + TEST_TBL, &select_op_error));
+  ASSERT_RAISES(IOError, session_error->ExecuteStatement("select * from " + TEST_TBL,
+                                                         &select_op_error));
   EXPECT_OK(session_error->Close());
 }
 
@@ -249,7 +417,7 @@ TEST(ServiceTest, TestConnect) {
   EXPECT_FALSE(service->IsConnected());
   EXPECT_OK(service->Close());
   unique_ptr<Session> session3;
-  EXPECT_ERROR(service->OpenSession(user, config, &session3));
+  ASSERT_RAISES(IOError, service->OpenSession(user, config, &session3));
   EXPECT_OK(session3->Close());
 
   // We should be able to call Close again without errors.
@@ -265,14 +433,17 @@ TEST(ServiceTest, TestFailedConnect) {
   unique_ptr<Service> service;
 
   string invalid_host = "does_not_exist";
-  EXPECT_ERROR(
-      Service::Connect(invalid_host, port, conn_timeout, protocol_version, &service));
+  ASSERT_RAISES(IOError, Service::Connect(invalid_host, port, conn_timeout,
+                                          protocol_version, &service));
 
   int invalid_port = -1;
-  EXPECT_ERROR(
-      Service::Connect(host, invalid_port, conn_timeout, protocol_version, &service));
+  ASSERT_RAISES(IOError, Service::Connect(host, invalid_port, conn_timeout,
+                                          protocol_version, &service));
 
   ProtocolVersion invalid_protocol_version = ProtocolVersion::PROTOCOL_V2;
-  EXPECT_ERROR(
-      Service::Connect(host, port, conn_timeout, invalid_protocol_version, &service));
+  ASSERT_RAISES(IOError, Service::Connect(host, port, conn_timeout,
+                                          invalid_protocol_version, &service));
 }
+
+}  // namespace hiveserver2
+}  // namespace arrow
